% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaguettiPlot.R
\name{SpaguettiPlot}
\alias{SpaguettiPlot}
\title{SpaguettiPlot function
This function will highlight one variable (NewDeaths, NewCases, CasesIncidence, DeathsIncidence) for one country against other countries}
\usage{
SpaguettiPlot(Country, Variable, AllOrSelection, ListCountriesToCompare, col)
}
\arguments{
\item{Country}{\itemize{
\item Country to highlight
}}

\item{Variable}{\itemize{
\item Variable can be NewCases,NewDeaths,NewCases_Rate,NewDeaths_Rate (will be developed further with other variables)
}}

\item{AllOrSelection}{\itemize{
\item Use 'All' if you want to compare the selected country to all other countries in EURO Region, use 'c(CountryA,CountryB,CountryC,...)' if you want to compare to some countries only
}}

\item{color}{\itemize{
\item Color to use to highlight the country Ex: '#000000'
}}
}
\description{
SpaguettiPlot function
This function will highlight one variable (NewDeaths, NewCases, CasesIncidence, DeathsIncidence) for one country against other countries
}
